2018-10-30 Version 10.13.0 'Dubnium' (LTS)
This release marks the transition of Node.js 10.x into Long Term
Support (LTS) with the codename 'Dubnium'. The 10.x release line
now moves in to "Active LTS" and will remain so until April 2020.
After that time it will move in to "Maintenance" until end of
life in April 2021.

Notable Changes:

There are no changes in this release aside from transitioning to LTS.