modules: introduce nodejs namespace
Up until now built-in modules shared a namepsace with all other modules on
the file system. This commit introduces a url based namespace for the
Node.js project to use for built-in modules.

From this point on all future modules adding by Node.js core will be exposed
in the `nodejs:` namespace. Legacy modules will exist on both the `nodejs:`
and global namespaces sharing the same singleton.

`require('nodejs:fs') === require('fs')`

As `worker_threads` is still an experimental feature it has been moved into the
`nodejs:` namespace.

Why `nodejs:` and node `@nodejs/`?

The `@nodejs/` namespace is already reserved on `npm` and there are plans to
use this namespace for modules we publish to `npm`. Making a distinction
between modules prepackaged in node and those being distributed externally
seems like a prudent choice.